% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithcv_of_lognorm_function.R
\name{arithcv_of_lognorm}
\alias{arithcv_of_lognorm}
\title{Arithmetic CV of lognormal distribution}
\usage{
arithcv_of_lognorm(logsd)
}
\arguments{
\item{logsd}{Lognormal standard deviation}
}
\value{
Arithmetic coefficient of variation of lognormal distribution
}
\description{
Calculation of the arithmetic coefficient of variation of a lognormal
distribution from function parameters
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{CV}
\keyword{SD}
\keyword{arithmetic}
\keyword{coefficient}
\keyword{deviation}
\keyword{lognormal}
\keyword{moments}
\keyword{of}
\keyword{standard}
\keyword{variation}
