% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmean_of_lognorm_function.R
\name{arithmean_of_lognorm}
\alias{arithmean_of_lognorm}
\title{Arithmetic mean of lognormal distribution}
\usage{
arithmean_of_lognorm(logmean, logsd)
}
\arguments{
\item{logmean}{Lognormal mean}

\item{logsd}{Lognormal standard deviation}
}
\value{
Arithmetic mean of lognormal distribution
}
\description{
Calculation of the arithmetic mean of a lognormal
distribution from function parameters
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{SD}
\keyword{arithmetic}
\keyword{average}
\keyword{deviation}
\keyword{lognormal}
\keyword{mean}
\keyword{moments}
\keyword{standard}
