% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithsd_of_lognorm_function.R
\name{arithsd_of_lognorm}
\alias{arithsd_of_lognorm}
\title{Arithmetic SD of lognormal distribution}
\usage{
arithsd_of_lognorm(logmean, logsd)
}
\arguments{
\item{logmean}{Lognormal mean}

\item{logsd}{Lognormal standard deviation}
}
\value{
Arithmetic standard deviation of lognormal distribution
}
\description{
Calculation of the arithmetic standard deviation of a lognormal
distribution from function parameters
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
