% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chmean_function.R
\name{chmean}
\alias{chmean}
\title{Contraharmonic mean}
\usage{
chmean(x, na.rm = T)
}
\arguments{
\item{x}{Input vector}

\item{na.rm}{Boolean to remove missing values}
}
\value{
Contraharmonic mean
}
\description{
Calculation of contraharmonic mean.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{average}
\keyword{contraharmonic}
\keyword{mean}
