% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgamma_mix_function.R
\name{dgamma_mix}
\alias{dgamma_mix}
\title{Density of mixed gamma distribution}
\usage{
dgamma_mix(x, shape, rate)
}
\arguments{
\item{x}{Input vector of values}

\item{shape}{Vector of shape parameters of all contributing gamma
distributions}

\item{rate}{Vector of rate parameters of all contributing gamma
distributions}
}
\value{
Vector of densities for each value in x
}
\description{
Probability density function for a mixture distribution of several gamma
distributions (all contributing with equal weights)
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{PDF}
\keyword{density}
\keyword{distribution}
\keyword{gamma}
\keyword{mixture}
\keyword{sum}
