% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispvar_function.R
\name{dispvar}
\alias{dispvar}
\title{Dispersion variance}
\usage{
dispvar(vgm, small.scale = 100, large.scale = 1000)
}
\arguments{
\item{vgm}{Variogram model (gstat) of a variable at small scale support}

\item{small.scale}{Side length of an area unit of the smaller scale}

\item{large.scale}{Side length of an area unit of the larger scale}
}
\value{
Dispersion variance
}
\description{
Calculation of dispersion variance between two scales. This only works if a
variogram model is available for the small scale. The dispersion variance
quantifies how much of the total variance of a variable at small scale is
happening inside the units of a larger units. You can also see it as the
proportion of variance being lost when upscaling from the small to the
large scale. The variance at large scale is the variance at small scale
minus the dispersion variance between the two scales.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
