% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_trees_with_spacing_function.R
\name{distribute_trees_with_spacing}
\alias{distribute_trees_with_spacing}
\title{Distribute trees with spacing}
\usage{
distribute_trees_with_spacing(
  d.vec,
  minx = 0,
  maxx = 100,
  miny = 0,
  maxy = 100,
  spacing.factor = 5
)
}
\arguments{
\item{d.vec}{Vector of tree stem diameters}

\item{minx}{Minimal X-coordinate of the area of interest}

\item{maxx}{Maximal X-coordinate of the area of interest}

\item{miny}{Minimal Y-coordinate of the area of interest}

\item{maxy}{Maximal Y-coordinate of the area of interest}

\item{spacing.factor}{Factor that specifies the minimal possible distance to the next neighbor tree as a function of tree stem diameter}
}
\value{
data.table of trees with positions
}
\description{
Distribute trees in space with a size dependent minimal distance
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
