% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmixture_AGB_from_white_shot_noise_function.R
\name{dmixture_AGB_from_white_shot_noise}
\alias{dmixture_AGB_from_white_shot_noise}
\title{Density of mixture distribution of aboveground biomass from white shot noise}
\usage{
dmixture_AGB_from_white_shot_noise(
  x,
  G.dist = "lnorm",
  G.par1,
  G.par2,
  m.alpha,
  m.lambda
)
}
\arguments{
\item{x}{Input vector of AGB values}

\item{G.dist}{Type of distribution of growth G ("lnorm" or "gamma")}

\item{G.par1}{Parameter 1 of the distribution of G (logmean or gamma shape)}

\item{G.par2}{Parameter 2 of the distribution of G (logsd or gamma rate)}

\item{m.alpha}{Mortality parameter 1: white shot noise mean intensity}

\item{m.lambda}{Mortality parameter 2: white shot noise mean frequency}
}
\value{
Vector of densities for each value in x
}
\description{
Probability density function for a mixture distribution of several gamma distributions
of aboveground biomass (AGB), which arise from biomass growth G and biomass mortality m
with G following a distribution and m being assumed as white shot noise. Under these
assumptions the distribution of AGB can be modelled as follows...
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{AGB}
\keyword{PDF}
\keyword{density}
\keyword{distribution}
\keyword{gain}
\keyword{gamma}
\keyword{growth}
\keyword{mixture}
\keyword{mortality}
\keyword{noise}
\keyword{shot}
\keyword{sum}
\keyword{white}
