% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmixture_lnorm_gumbel_function.R
\name{dmixture_lnorm_gumbel}
\alias{dmixture_lnorm_gumbel}
\title{Density of mixture distribution of lognormal and gumbel distribution}
\usage{
dmixture_lnorm_gumbel(x, logmean, logsd, gumbel.alpha, gumbel.scale,
  w.gumbel = 0.05)
}
\arguments{
\item{x}{Input vector of values}

\item{logmean}{Logmean parameter of the lognormal distribution}

\item{logsd}{Logsd parameter of the lognormal distribution}

\item{gumbel.alpha}{Location parameter of the gumbel distribution}

\item{gumbel.scale}{Scale parameter of the gumbel distribution}

\item{w.gumbel}{Weight of the gumbel distribution within the mixture (weight of lognormal is 1-w.gumbel)}
}
\value{
Vector of densities for each value in x
}
\description{
Probability density function for a mixture distribution composed of one lognormal and
one gumbel distribution. The gumbel distribution is suited to model the extreme values
in a heavy-tail dataset, while the lognormal models the regular values.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{PDF}
\keyword{density}
\keyword{distribution}
\keyword{extreme}
\keyword{gumbel}
\keyword{lognormal}
\keyword{mixture}
\keyword{outlier}
\keyword{sum}
\keyword{theory}
\keyword{value}
