% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_mean_from_shape_rate_function.R
\name{gamma_mean_from_shape_rate}
\alias{gamma_mean_from_shape_rate}
\title{Gamma mean from shape and rate parameter}
\usage{
gamma_mean_from_shape_rate(shape, rate)
}
\arguments{
\item{shape}{parameter of gamma distribution}

\item{rate}{parameter of gamma distribution}
}
\value{
Mean of gamma distribution
}
\description{
Calculation of the mean of a gamma distribution from its shape and rate
parameters
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
