% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_par_from_gen_gamma_par_function.R
\name{gamma_par_from_gen_gamma_par}
\alias{gamma_par_from_gen_gamma_par}
\title{Gamma parameters from generalized gamma parameters and transformation power
law coefficients}
\usage{
gamma_par_from_gen_gamma_par(
  pl.fac,
  pl.exp,
  gg.a,
  gg.d,
  gg.p,
  gg.mu,
  gg.sigma,
  gg.Q,
  flexsurv.form = F
)
}
\arguments{
\item{pl.fac}{Power law prefactor}

\item{pl.exp}{Power law exponent}

\item{gg.a}{GG parameter}

\item{gg.d}{GG parameter}

\item{gg.p}{GG parameter}

\item{gg.mu}{GG parameter (flexsurv form)}

\item{gg.sigma}{GG parameter (flexsurv form)}

\item{gg.Q}{GG parameter (flexsurv form)}

\item{flexsurv.form}{Switch whether to use the default GG parameter set
(a, d, p) or the flexsurv GG parameter set (mu, sigma, Q)}
}
\value{
Vector containing gamma parameters shape, scale, rate as well as
descriptive statistics mean, var, sd and skewness of the gamma distribution
}
\description{
If a variable x follows a gamma distribution the dependent variable y, which
is related to x via a power law relationship y=a*x^b follows a generalized
gamma distribution (GG). The parameters of the gamma distribution can be
calculated from the input GG parameters and the power law coefficients.
Two alternative formulations for the GG exist in the literature,
one of which uses parameters named a, d, p while the other one uses
mu, sigma, Q (e.g., R package flexsurv). Both parameter sets are
accepted by this function.
For parameter conversion see also:
https://en.wikipedia.org/wiki/Generalized_gamma_distribution
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
