% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_rate_from_mean_var_function.R
\name{gamma_rate_from_mean_var}
\alias{gamma_rate_from_mean_var}
\title{Gamma rate parameter from mean and variance}
\usage{
gamma_rate_from_mean_var(mean, var)
}
\arguments{
\item{mean}{of gamma distribution}

\item{variance}{of gamma distribution}
}
\value{
Rate parameter of gamma distribution
}
\description{
Calculation of the rate parameter of a gamma distribution from its
mean and variance
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
