% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_var_from_shape_rate_function.R
\name{gamma_var_from_shape_rate}
\alias{gamma_var_from_shape_rate}
\title{Gamma variance from shape and rate parameter}
\usage{
gamma_var_from_shape_rate(shape, rate)
}
\arguments{
\item{shape}{parameter of gamma distribution}

\item{rate}{parameter of gamma distribution}
}
\value{
Variance of gamma distribution
}
\description{
Calculation of the variance of a gamma distribution from its
shape and rate parameters
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
