% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_convergence_factor_function.R
\name{Gaussian_convergence_factor}
\alias{Gaussian_convergence_factor}
\title{Scaling factor where distribution converges to Gaussian}
\usage{
Gaussian_convergence_factor(vec, alpha = 0.05, BerryEsseen.const = 0.4748)
}
\arguments{
\item{vec}{Input data vector}

\item{alpha}{Significance level for the Kolmogorov-Smirnov statistic}

\item{BerryEsseen.const}{Berry-Esseen constant (there are different values found in the literature)}
}
\value{
Scale factor between the current scale and the scale of Gaussian convergence (number of units that have to be aggregated, respectively)
}
\description{
According to the central limit theorem distributions of aggregated variables
converge to a Gaussian shape with increasing aggregation. This function
calculates the aggregation factor between the current scale and the scale of
convergence for a given data vector. It is based on the Berry-Esseen theorem,
which provides an upper bound for the distance between standard normal CDF
and z-transformed empirical CDF (based on the 3rd moment). This distance
between CDFs is the Kolmogorov-Smirnov statistic D and for a desired
significance level alpha the critical value of D can be calculated.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
