% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gamma_par_from_gamma_par_function.R
\name{gen_gamma_par_from_gamma_par}
\alias{gen_gamma_par_from_gamma_par}
\title{Generalized gamma parameters from gamma parameters and transformation power
law coefficients}
\usage{
gen_gamma_par_from_gamma_par(pl.fac, pl.exp, g.shape, g.rate)
}
\arguments{
\item{pl.fac}{Power law prefactor}

\item{pl.exp}{Power law exponent}

\item{g.shape}{Gamma shape parameter}

\item{g.rate}{Gamma rate parameter}
}
\value{
Vector containing GG parameters a, d, p and mu, sigma, Q as well as
descriptive statistics mean, var and sd of the GG
}
\description{
If a variable x follows a gamma distribution the dependent variable y, which
is related to x via a power law relationship y=a*x^b follows a generalized
gamma distribution (GG). The parameters of the GG can be calculated from the
input gamma parameters and the power law coefficients. Two alternative
formulations for the GG exist in the literature, one of which uses parameters
named a, d, p while the other one uses mu, sigma, Q (e.g., R package
flexsurv). Both parameter sets are calculated by this function.
For parameter conversion see also:
https://en.wikipedia.org/wiki/Generalized_gamma_distribution
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
