% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmean_function.R
\name{gmean}
\alias{gmean}
\title{Geometric mean}
\usage{
gmean(x, na.rm = T, zero.rm = F)
}
\arguments{
\item{x}{Input vector}

\item{na.rm}{Boolean to remove missing values}
}
\value{
Geometric mean
}
\description{
Calculation of geometric mean.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
