% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmean_function.R
\name{hmean}
\alias{hmean}
\title{Harmonic mean}
\usage{
hmean(x, na.rm = T)
}
\arguments{
\item{x}{Input vector}

\item{na.rm}{Boolean to remove missing values}
}
\value{
Harmonic mean
}
\description{
Calculation of harmonic mean.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{average}
\keyword{harmonic}
\keyword{mean}
