% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_of_gumbel_function.R
\name{loc_of_gumbel}
\alias{loc_of_gumbel}
\title{Location parameter of gumbel distribution}
\usage{
loc_of_gumbel(mean, var = NA, scale = NA)
}
\arguments{
\item{mean}{Mean of the gumbel distribution}

\item{var}{Variance of the gumbel distribution}

\item{scale}{Scale parameter of the gumbel distribution}
}
\value{
Location of gumbel distribution
}
\description{
Calculation of the location parameter (sometimes called alpha) of a gumbel
distribution from its mean and either variance or scale parameter (don't
provide both)
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
