% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmean_of_lognorm_function.R
\name{logmean_of_lognorm}
\alias{logmean_of_lognorm}
\title{Lognormal mean of lognormal distribution}
\usage{
logmean_of_lognorm(arithmean, arithsd)
}
\arguments{
\item{arithmean}{Arithmetic mean}

\item{arithsd}{Arithmetic standard deviation}
}
\value{
Logarithmic standard deviation of lognormal distribution (sigma
parameter)
}
\description{
Calculation of the logarithmic mean parameter (mu)
of a lognormal distribution from distribution moments
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
