% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsd_of_lognorm_function.R
\name{logsd_of_lognorm}
\alias{logsd_of_lognorm}
\title{Lognormal SD of lognormal distribution}
\usage{
logsd_of_lognorm(arithmean, arithsd)
}
\arguments{
\item{arithmean}{Arithmetic mean}

\item{arithsd}{Arithmetic standard deviation}
}
\value{
Logarithmic standard deviation of lognormal distribution (sigma parameter)
}
\description{
Calculation of the logarithmic standard deviation parameter (sigma)
of a lognormal distribution from distribution moments
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
