% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_gumbel_function.R
\name{mean_of_gumbel}
\alias{mean_of_gumbel}
\title{Mean of gumbel distribution}
\usage{
mean_of_gumbel(loc, scale)
}
\arguments{
\item{loc}{Location parameter of the gumbel distribution}

\item{scale}{Scale parameter of the gumbel distribution}
}
\value{
Arithmetic mean of gumbel distribution
}
\description{
Calculation of the arithmetic mean of a gumbel
distribution from function parameters
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{arithmetic}
\keyword{average}
\keyword{extreme}
\keyword{gumbel}
\keyword{location}
\keyword{mean}
\keyword{moments}
\keyword{scale}
\keyword{value}
