% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_mixture_function.R
\name{mean_of_mixture}
\alias{mean_of_mixture}
\title{Mean of mixture distribution}
\usage{
mean_of_mixture(means, weights = NA)
}
\arguments{
\item{means}{Vector of means of all contributing distributions}

\item{weights}{Vector of weights of all contributing distributions}
}
\value{
Arithmetic mean of mixture distribution
}
\description{
Calculation of the mean of a mixture distribution from means of all contributing distributions
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
