% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_from_quantiles_function.R
\name{pdf_from_quantiles}
\alias{pdf_from_quantiles}
\title{Reconstruct probability density function from quantiles}
\usage{
pdf_from_quantiles(q.vec, p.vec, x.vec)
}
\arguments{
\item{q.vec}{Vector of quantiles (quantile function / inverse CDF; has to be
sorted in increasing order)}

\item{p.vec}{Vector of probabilities (values between 0 and 1 providing the
probability associated with each element of q.vec)}

\item{x.vec}{Vector of values for which the densities should be returned}
}
\value{
Vector of probability densities for each element in x.vec
}
\description{
Reconstruct a PDF based on a set of quantiles of a distribution
or the quantile function (inverse cumulative distribution function ICDF)
}
\examples{
my.r.vec <- rnorm(n=50, mean=10, sd=6)
my.q.vec <- sort(my.r.vec)
my.p.vec <- pnorm(my.q.vec, mean=10, sd=6)
plot(my.q.vec ~ my.p.vec, type="l")

my.x.vec <- -20:40
my.pdf.vec <- pdf_from_quantiles(q.vec=my.q.vec, p.vec=my.p.vec,
x.vec=my.x.vec)

hist(my.r.vec, freq=F, xlim=c(-20, 40))
lines(my.pdf.vec ~ my.x.vec, col="red")
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
