% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_transform_vec_function.R
\name{rank_transform_vec}
\alias{rank_transform_vec}
\title{Rank transformation based on a vector of target values}
\usage{
rank_transform_vec(x, y)
}
\arguments{
\item{x}{Input vector}

\item{y}{Target vector}
}
\value{
Vector of transformed values
}
\description{
Rank transform a dataset, i.e., for each input value provide an output value
from a given target vector. The target vector will be rearranged such that
the ranking of values corresponds the ranking of the input vector values.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
