% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdist_function.R
\name{rdist}
\alias{rdist}
\title{Random numbers from distribution}
\usage{
rdist(n, x.vec, y.vec)
}
\arguments{
\item{n}{Number of samples that should be drawn from the distribution}

\item{x.vec}{Vector defining the values of the distribution}

\item{y.vec}{Vector defining the probability density or frequencies of the x values}
}
\value{
Vector of random numbers
}
\description{
Random number generation from arbitrary distribution defined by
a value vector x and a probability or frequency vector y. The function
uses inverse transform sampling.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{CDF}
\keyword{arbitrary}
\keyword{custom}
\keyword{distribution}
\keyword{generator}
\keyword{inverse}
\keyword{number}
\keyword{random}
\keyword{sampling}
\keyword{transform}
\keyword{uniform}
