% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_par_function.R
\name{rescale_par}
\alias{rescale_par}
\title{Rescale parameter}
\usage{
rescale_par(x, x.scale, y.scale, scaling.exp = -0.5)
}
\arguments{
\item{x}{Input value, e.g., known distribution parameter at a certain scale (reference scale)}

\item{x.scale}{Side length of a measurement unit of the input (reference scale)}

\item{y.scale}{Side length of a measurement unit of the output (target scale)}

\item{scaling.exp}{Scaling exponent}
}
\value{
Parameter value at the target scale
}
\description{
Rescaling of distribution parameters (e.g., standard deviation) based on a
scaling exponent, i.e., a log-log-linear relationship (power law) between the area ratio of
input and output scales and the ratio of the known input parameter value and the unknown
output parameter value.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
