% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_of_gumbel_function.R
\name{scale_of_gumbel}
\alias{scale_of_gumbel}
\title{Scale of gumbel distribution}
\usage{
scale_of_gumbel(var)
}
\arguments{
\item{var}{Variance of the gumbel distribution}
}
\value{
Scale of gumbel distribution
}
\description{
Calculation of the scale parameter (sometimes called beta) of a gumbel
distribution from its variance
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{SD}
\keyword{arithmetic}
\keyword{deviation}
\keyword{extreme}
\keyword{gumbel}
\keyword{location}
\keyword{moments}
\keyword{scale}
\keyword{standard}
\keyword{value}
\keyword{variance}
