% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_distr_function.R
\name{transform_distr}
\alias{transform_distr}
\title{Transform a distribution with a function}
\usage{
transform_distr(
  x.vec,
  xy.func = function(x) {     1 * x },
  xsdy.func = function(x) {     1 * x },
  n.steps = 1000
)
}
\arguments{
\item{x.vec}{Vector defining the values of the distribution}

\item{n}{Number of samples that should be drawn from the distribution}

\item{y.vec}{Vector defining the probability density or frequencies of the x values}
}
\value{
Vector of random numbers
}
\description{
The PDF of an independent variable x is transformed into the PDF of
a dependent variable y, based on their relationship y=f\(x\)
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
