% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pdf_with_function_function.R
\name{transform_pdf_with_function}
\alias{transform_pdf_with_function}
\title{Transform a probability density function with a function}
\usage{
transform_pdf_with_function(
  x.vec,
  x.d.vec,
  xy.func = function(x) {     1 * x },
  n.steps = 1000
)
}
\arguments{
\item{x.vec}{Vector of values of the independent variable}

\item{x.d.vec}{Vector of probability densities of each x value}

\item{xy.func}{Function which relates x to y; Default is y=x, i.e. as an R
function function(x){1*x}}

\item{n.steps}{Number of discrete steps between which the PDF is
approximated via linear interpolation during the transformation}
}
\value{
List containing 1) y values for each input x value, 2) x values for
all n steps, 3) y values for all n steps (of x) and 4) PDF values for all n
steps, as well as the functions for the PDFs of 5) x and 6) y as approxfun
functions
}
\description{
The PDF of an independent variable x is transformed into the PDF of
a dependent variable y, based on their relationship y=f(x).
}
\examples{
# Simulate data from a lognormal distribution
my.x.vec <- rlnorm(200, 3.2, 0.3)
my.x.d.vec <- dlnorm(my.x.vec, 3.2, 0.3)
hist(my.x.vec, breaks=30, freq=F)
curve(dlnorm(x, 3.2, 0.3), add=T, col="blue")
# Define a power law relationship y=0.1*x^1.8
my.xy.func <- function(x){0.1*x^1.8}
# Apply the transformation
transform.list <- transform_pdf_with_function(x.vec=my.x.vec,
x.d.vec=my.x.d.vec, xy.func=my.xy.func, n.steps=1000)
hist(transform.list$y)
plot(transform.list$y ~ my.x.vec)
plot(transform.list$pdf.steps ~ transform.list$x.steps)
plot(transform.list$pdf.steps ~ transform.list$y.steps)
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
