% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_of_mixture_function.R
\name{var_of_mixture}
\alias{var_of_mixture}
\title{Variance of mixture distribution}
\usage{
var_of_mixture(means, vars, weights = NA)
}
\arguments{
\item{means}{Vector of means of all contributing distributions}

\item{vars}{Vector of variances of all contributing distributions}

\item{weights}{Vector of weights of all contributing distributions}
}
\value{
Variance of mixture distribution
}
\description{
Calculation of the variance of a mixture distribution from means and variances of all contributing distributions
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
