% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_of_gumbel_function.R
\name{var_of_gumbel}
\alias{var_of_gumbel}
\title{Variance of gumbel distribution}
\usage{
var_of_gumbel(scale)
}
\arguments{
\item{scale}{Scale parameter of the gumbel distribution}
}
\value{
Variance of gumbel distribution
}
\description{
Calculation of the variance of a gumbel
distribution from function parameter
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{SD}
\keyword{arithmetic}
\keyword{deviation}
\keyword{extreme}
\keyword{gumbel}
\keyword{location}
\keyword{moments}
\keyword{scale}
\keyword{standard}
\keyword{value}
\keyword{variance}
